Ignore 'UnknownTopicOrPartition' error
Prior to this fix, if a topic is deleted while bruke is running
brod_consumer exit with 'normal' reason and no restart by supervisor.
then brucke_subscriber will keep trying re-subscribe in a delay-retry
dead-loop.

This fix is to silently discard the 'UnknownTopicOrPartition' error,
and keep the group balanced as much as possible.

Hopefully ops will notice it (from health-check or lagging monitor)
and make a config change.